+++ tests/misc.spec.ts
@@ -5,26 +5,16 @@ import { parser } from "src/processor/parser"
 import { replacer } from "src/replacer"
 
 const content = `
-    <div className={twg(
-        "size-92 relative grid place-items-center px-4 py-2",
+    <div className={cn(
+        "multiple classes",
         {
-            before: isTernary === "foo" ? [
-                "absolute inset-0 bg-red-500",
-                {
-                    hover: isTernary ? "bg-blue-500 text-yellow-500" : [
-                        "bg-blue-500 text-yellow-500",
-                        isAndOr && "border-2 border-white"
-                    ]
-                }
-            ] : [
-                "fixed inset-0 bg-yellow-500",
-            ],
-            "aria-expanded": "bg-red-500 text-yellow-500",
+            mod1: ["class", "other classes"],
+            mod2: ["class", { "additional-mod": "other classes" }]
         }
     )} />
 ` as string
 
-const transformedContent = liteReplacer()(content)
+const transformedContent = replacer({ callee: [] })(content)
 
 console.log("1: ", transformedContent)